[src] Fix a lot of other cases of IntPtr -> NativeHandle.
This is just fallout from the CFArray/NSArray in the previous commit.

+++ src/CoreText/CTFont.cs
@@ -2356,7 +2356,7 @@ namespace CoreText {
 			if (cfArrayRef == IntPtr.Zero)
 				return Array.Empty<CTFontTable> ();
 			return NSArray.ArrayFromHandle (cfArrayRef, v => {
-					return (CTFontTable) (uint) v;
+					return (CTFontTable) (uint) (IntPtr) v;
 			}, true);
 		}
